Update install.sh
Ubuntu 22.04 on WSL2 does not come preinstalled with `pdftoppm`, which causes the script to fail, therefore I added `poppler-utils` which provides it.
It also does not come with `bc`; however this does not cause the script to fail, but display a "command not found" at the end. I included it to enable full functionality of the script.